feat: integrate frontend and backend for delete team and fetch teams based on user id

+++ backend/team-micro-service/src/main/java/com/tma/team_micro_service/service/implementation/TeamServiceImplementation.java
@@ -41,9 +41,20 @@ public class TeamServiceImplementation implements TeamService {
       Team savedTeam = teamRepository.save(team);
 
       log.info("Saved team and calling user feign client to add team id to user");
+      
+      
+      log.info("INside create team function");
+      
+      log.info("Team Id: {}", savedTeam.getTeamId());
+      log.info("User Id: {}", userId);
 
       userFeignClient.addTeamToUser(savedTeam.getTeamId(), userId, bearerToken);
-
+      
+      log.info("After feign client call");
+      
+      log.info("Team Id: {}", savedTeam.getTeamId());
+      log.info("User Id: {}", userId);
+      
       log.info("Success");
 
       return savedTeam;
@@ -83,9 +94,13 @@ public class TeamServiceImplementation implements TeamService {
   }
 
   @Override
-  public void deleteTeam(UUID teamId) {
-
-    teamRepository.deleteById(teamId);
+  public void deleteTeam(UUID teamId, UUID userId, HttpServletRequest request) {
+    String bearerToken = request.getHeader("Authorization");
+    
+    if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
+      teamRepository.deleteById(teamId);
+      userFeignClient.removeUserFromTeam(userId, teamId, bearerToken);
+    }
   }
 
   @Override
@@ -99,9 +114,4 @@ public class TeamServiceImplementation implements TeamService {
 
     return team.getUserIds();
   }
- 
-  
-  
-  
-  
 }
